+++ src/form/formula_gen.cpp
@@ -572,9 +572,15 @@ bool FormulaGenerator::generateSingle(const Program& p) {
     Log::get().debug("Processed post-loop: " + formula.toString());
   }
 
-  // resolve simple recursions
-  FormulaSimplify::replaceTrivialRecursions(formula);
-  Log::get().debug("Resolved simple recursions: " + formula.toString());
+  // replace arithmetic progressions
+  if (FormulaSimplify::replaceArithmeticProgressions(formula)) {
+    Log::get().debug("Replaced arithmetic progressions: " + formula.toString());
+  }
+
+  // replace geometric progressions
+  if (FormulaSimplify::replaceGeometricProgressions(formula)) {
+    Log::get().debug("Replaced geometric progressions: " + formula.toString());
+  }
 
   // resolve simple functions
   FormulaSimplify::resolveSimpleFunctions(formula);
@@ -768,6 +774,11 @@ bool FormulaGenerator::generate(const Program& p, int64_t id, Formula& result,
 
   // replace simple references to recursive functions
   FormulaSimplify::replaceSimpleRecursiveRefs(result);
+
+  // replace arithmetic & geometric progressions
+  // TODO: check if this is really needed here
+  FormulaSimplify::replaceArithmeticProgressions(formula);
+  FormulaSimplify::replaceGeometricProgressions(result);
 
   // replace functions A000142(n) by n! in all formula definitions
   const auto factorialSeqName = FACTORIAL_SEQ_ID.string();